Fase 1: perfil/endereços/dispositivos/notificações + seed dev + correções
- ProfileController: perfil, CRUD de endereços (protege endereço usado em pedido), registro de token FCM, inbox
- seed-dev.sql: catálogo sintético (6 produtos, promoção, banner) para desenvolvimento
- Corrige inferência de tipo no upsert do carrinho (coalesce($4::numeric,0))

E2E validado: login -> endereço -> carrinho (promoção aplicada) -> pedido PIX ->
outbox -> mock ERP -> FATURADO -> EM_SEPARACAO -> SAIU_ENTREGA, com PIX copia-e-cola,
NF e log de integração gravados.

+++ apps/api/src/orders/orders.service.ts
@@ -54,7 +54,7 @@ export class OrdersService {
       );
       await pool.query(
         `insert into carrinho_itens (carrinho_id, produto_id, quantidade, preco_unit_snapshot)
-         values ($1,$2,$3, coalesce($4,0))
+         values ($1,$2,$3, coalesce($4::numeric,0))
          on conflict (carrinho_id, produto_id) do update set quantidade = $3`,
         [id, produtoId, quantidade, preco.rows[0]?.preco ?? 0],
       );
